move libs/projectAccess.mjs to utils/
